pyproject.toml
[project]
name = "udigest"
version = "0.1.0"
description = "A digest of Hugging Face papers."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.4.0",
    "beautifulsoup4>=4.14.3",
    "ddgs>=9.10.0",
    "dspy>=3.1.2",
    "jinja2>=3.1.6",
    "tqdm>=4.67.2",
    "typer-slim>=0.21.1",
    "wikipedia>=1.4.0",
]
